+++ lib/parse/Grammar.go
@@ -6,6 +6,7 @@ type Grammar struct {
 	writeRules            *GrammarRules
 	tokenizer             *Tokenizer
 	morphologicalAnalyzer *MorphologicalAnalyzer
+	texts				  map[string]string
 }
 
 func NewGrammar(locale string) Grammar {
@@ -15,11 +16,25 @@ func NewGrammar(locale string) Grammar {
 		writeRules:            NewGrammarRules(),
 		tokenizer:             NewTokenizer(DefaultTokenizerExpression),
 		morphologicalAnalyzer: nil,
+		texts:				   map[string]string{},
 	}
 }
 
 func (grammar *Grammar) SetTokenizer(tokenizer *Tokenizer) {
 	grammar.tokenizer = tokenizer
+}
+
+func (grammar *Grammar) SetTexts(texts map[string]string) {
+	grammar.texts = texts
+}
+
+func (grammar *Grammar) GetText(text string) string {
+	translation, found := grammar.texts[text]
+	if found {
+		return translation
+	} else {
+		return text
+	}
 }
 
 func (grammar *Grammar) GetTokenizer() *Tokenizer {